Add Similar Roles section to job detail pages with broad fallback
Convert /jobs/[id] from a client component to a server component + client
shell (JobDetailClient) so the similar-roles cards land in the initial
HTML response. Expired pages are the main beneficiary: they keep their
SEO equity (status = approved/expired stays indexable, per the recent
sitemap/layout work) but now convert lost-intent traffic into clicks on
live roles instead of dead-ending.

Changes:
- New lib/jobs/similar-jobs.ts — 3-step query ladder (same category +
  location -> same category any location -> related categories) plus a
  broad random fallback that fires only when the narrow ladder returns
  zero. Shuffles client-side so different expired URLs show different
  cards, which avoids duplicate-content signals. Filters status =
  'approved' so expired roles never appear as similar.
- New components/jobs/SimilarRoles.tsx — pure presentational, accepts
  pre-fetched SimilarJobsResult as a prop. Swaps CTA + URL when the
  fallback is 'broad' (View all AI roles -> /jobs) vs 'similar' (uses
  existing getSimilarRolesUrl).
- New lib/jobs/get-job-by-id.ts — shared React.cache() helper so the
  layout's generateMetadata, the layout's JSON-LD renderer, and the
  new server page.tsx all share a single DB round-trip per request.
- app/jobs/[id]/layout.tsx — uses the shared getJobById.
- app/jobs/[id]/page.tsx — rewritten as a thin async server component
  that fetches the job and similar roles, then renders JobDetailClient.
- app/jobs/[id]/JobDetailClient.tsx — extracted interactive shell that
  keeps all auth/apply/save/modal state; accepts initialJob and
  similarRoles as props (no more client-side job fetch, no loading
  spinner for initial job data).

Layouts: active gets 3 cards in a full-width strip below the 2-col grid;
expired gets 4 cards full-width between the header row and description
row, so the user sees hiring roles immediately after confirming the job
is expired. Mobile stacks everything vertically in document order.

Verified locally: lint + tsc pass; server-rendered expired page source
contains the four cards inline with no reliance on JS execution, no
JobPosting JSON-LD on expired (per Google Jobs policy), and fallback =
'similar' on a Machine Learning Engineer page with matches in related
categories.
